requirements.txt
Markdown
python-frontmatter
Jinja2
Pygments